@@ -1,6 +1,6 @@
 module github.com/domonda/go-function/cli
 
-go 1.22
+go 1.23
 
 replace github.com/domonda/go-function => ../
 
@@ -9,7 +9,6 @@ require github.com/domonda/go-function v0.0.0-00010101000000-000000000000 // rep
 require (
 	github.com/fatih/color v1.17.0
 	github.com/posener/complete/v2 v2.1.0
-	golang.org/x/exp v0.0.0-20240525044651-4c93da0ed11d
 )
 
 require (
@@ -18,5 +17,5 @@ require (
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/posener/script v1.2.0 // indirect
 	github.com/ungerik/go-httpx v0.0.0-20240110134719-544aadceddba // indirect
-	golang.org/x/sys v0.20.0 // indirect
+	golang.org/x/sys v0.24.0 // indirect
 )
