@@ -1,15 +1,15 @@
 module github.com/domonda/go-function/htmlform
 
-go 1.22
+go 1.23
 
 replace github.com/domonda/go-function => ..
 
 require github.com/domonda/go-function v0.0.0-00010101000000-000000000000 // replaced
 
 require (
-	github.com/domonda/go-types v0.0.0-20240514083716-b9f8b57b701d
-	github.com/domonda/golog v0.0.0-20240413045622-e718d2ebc935
-	github.com/ungerik/go-fs v0.0.0-20240517093852-5c045d7130d7
+	github.com/domonda/go-types v0.0.0-20240807125919-3d418e4a028b
+	github.com/domonda/golog v0.0.0-20240805081224-f870d6f12732
+	github.com/ungerik/go-fs v0.0.0-20240702143946-3ecb6733945d
 	github.com/ungerik/go-httpx v0.0.0-20240110134719-544aadceddba
 )
 
@@ -20,8 +20,8 @@ require (
 	github.com/h2non/filetype v1.1.3 // indirect
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/mattn/go-runewidth v0.0.15 // indirect
+	github.com/mattn/go-runewidth v0.0.16 // indirect
 	github.com/muesli/termenv v0.15.2 // indirect
 	github.com/rivo/uniseg v0.4.7 // indirect
-	golang.org/x/sys v0.20.0 // indirect
+	golang.org/x/sys v0.24.0 // indirect
 )
